Resolve only the rootPath within dispatch_once in bugsnagPath()
Revert to how rootPath is resolved in upstream.

`rootPath` is now resolved within a dispatch_once block, then if `subdirectory` is specified it gets appended and that is returned instead.

It's important that when subdirectory is provided it is not resolved within the dispatch_once otherwise it will not resolved correctly if different configurations are created in the same processed (e.g. when batch uploading crash reports from different processes).